different type of trees and year of planting
Instead of choosing plant at year 0 or not, the strategies now consist in choosing the type of tree and year of planting. The user can move a slider to observe how the planting changes depending on the year. Clicking on a land parcel now prevents planting before a given year.