call host "backend" instead of localhost, the backend is somewhat functional
in the backend initialization, do not use an outer future i.e. it is now blocking because opening files in a future inside a future causes issues

notif now has a default empty string message parameter

the backend is code is now treated as read-only

TODO: save user Elicitor upload files in a new writeable folder e.g. in /tmp